Migrate to ASP.NET Core 6, New Http Request, Better ENV handling, Rewrite Twitch.cs
Deleted Startup.cs because Startup.cs and Program.cs are together in 6.0.

Changed all values which are needed for 6.0 in project files

Changed Webrequest to HttpClient now because it is recommended from Microsoft.

Rewritten TwitchController.cs and deleted twitch.cs. Everything is now in the controller. (Not finished yet. Needs some refactoring.)

Fixed bug where Twitch Website Section would crash the site when TwitchUserList is null.

We handle the environment better now. Clear line between Production and Development. Changed options in the build command. Also Changed the output directory for release so it's not in Debug anymore.