Refactored TwitchController.cs
We removed the Json file which was only for the TwitchController and instead put it in the TwitchModel class. It makes it more clear to read. We also changed all unnecessary 'Twitch' words in variable names because it's clear now that it is only about Twitch since we store it now in a Twitch class.